+++ scripts/approximations/main.py
@@ -23,19 +23,17 @@ def main():
     # end of the interval to calculate the approximation on
     x_end = 1
     # number of paramters to use for the approximations.
-    num_parameters = 6
+    num_parameters = 12
 
     # number of (x,y) coordinates used to plot the resulting approximation.
-    num_points_plot = 100000
-    # number of (x,y) coordinates used to plot the
-    # actual function that is evenly spaced on the X-axis.
-    num_points_plot_accurate = 50000
+    num_points_plot = 10000
+
     # function to approximate
     approximated_fn = lambda x: math.e**x
 
     # flag controlling whether to plot each approximation.
     # Plots if true.
-    shouldPlotApproximations = False
+    shouldPlotApproximations = True
 
     # flag controlling whether to compute max error for each approximation
     # given the equally spaced x coordinates.
@@ -45,59 +43,14 @@ def main():
     #####################
     # 2. Approximations 
 
-    # 2.1. Equispaced Polynomial Approximation
-    coefficients_equispaced_poly = approximations.equispaced_poly_approx(approximated_fn, x_start, x_end, num_parameters)
-    
-    # 2.2. Chebyshev Polynomial Approximation
-    coefficients_chebyshev_poly = approximations.chebyshev_poly_approx(approximated_fn, x_start, x_end, num_parameters)
-    
-    # 2.3. Chebyshev Rational Approximation
-    numerator_coefficients_chebyshev_rational, denominator_coefficients_chebyshev_rational = approximations.chebyshev_rational_approx(approximated_fn, x_start, x_end, num_parameters)
-
-    # 2.4. Actual With Large Number of Coordinates (evenly spaced on the X-axis)
-    x_accurate = np.linspace(x_start, x_end, num_points_plot_accurate)
-
-    #######################################
-    # 3. Compute (x,y) Coordinates To Plot
+    y_eqispaced_poly, y_chebyshev_poly, y_chebyshev_rational, y_actual = approximations.approximate_all_with_num_parameters(approximated_fn, x_start, x_end, num_parameters, num_points_plot)
 
     # Equispaced x coordinates to be used for plotting every approximation.
     plot_nodes_x = np.linspace(x_start, x_end, num_points_plot)
 
-    # 3.1 Equispaced Polynomial Approximation
-    plot_nodes_y_eqispaced_poly = polynomial.evaluate(plot_nodes_x, coefficients_equispaced_poly)
-
-    # 3.2 Chebyshev Polynomial Approximation
-    plot_nodes_y_chebyshev_poly = polynomial.evaluate(plot_nodes_x, coefficients_chebyshev_poly)
-
-    # 3.3 Chebyshev Rational Approximation
-    y_chebyshev_rational = rational.evaluate(plot_nodes_x, numerator_coefficients_chebyshev_rational.tolist(), denominator_coefficients_chebyshev_rational.tolist())
-
-    # 3.4 Actual With Large Number of Coordinate (evenly spaced on the X-axis)
-    y_accurate = approximated_fn(x_accurate)
 
     #############################
-    # 4. Plot Every Approximation
-
-    if shouldPlotApproximations:
-        # 4.1 Equispaced Polynomial Approximation
-        plt.plot(plot_nodes_x, plot_nodes_y_eqispaced_poly, label="Equispaced Poly")
-
-        # 4.2 Chebyshev Polynomial Approximation
-        plt.plot(plot_nodes_x,plot_nodes_y_chebyshev_poly, label="Chebyshev Poly")
-
-        # 4.3 Chebyshev Rational Approximation
-        plt.plot(plot_nodes_x,y_chebyshev_rational, label="Chebyshev Rational")
-
-        # 4.4 Actual With Large Number of Coordinates (evenly spaced on the X-axis)
-        plt.plot(x_accurate,y_accurate, label=F"Actual - {num_points_plot_accurate} evenly spaced points")
-
-        plt.legend(loc="upper left")
-        plt.grid(True)
-        plt.title(f"Appproximation of e^x on [{x_start}, {x_end}] with {num_parameters} parameters")
-        plt.show()
-
-    #############################
-    # 5. Compute Errors
+    # 4. Compute Errors
 
     if shouldComputeErrorDelta:
         print(F"\n\nMax Error on [{x_start}, {x_end}]")
@@ -106,17 +59,38 @@ def main():
 
         plot_nodes_y_actual = approximated_fn(plot_nodes_x)
 
-        # 5.1 Equispaced Polynomial Approximation
-        delta_eqispaced_poly = np.abs(plot_nodes_y_eqispaced_poly - plot_nodes_y_actual)
+        # 4.1 Equispaced Polynomial Approximation
+        delta_eqispaced_poly = np.abs(y_eqispaced_poly - plot_nodes_y_actual)
         print(F"Equispaced Poly: {np.amax(delta_eqispaced_poly)}")
 
-        # 5.2 Chebyshev Polynomial Approximation
-        delta_chebyshev_poly = np.abs(plot_nodes_y_chebyshev_poly - plot_nodes_y_actual)
+        # 4.2 Chebyshev Polynomial Approximation
+        delta_chebyshev_poly = np.abs(y_chebyshev_poly - plot_nodes_y_actual)
         print(F"Chebyshev Poly: {np.amax(delta_chebyshev_poly)}")
 
-        # 5.3 Chebyshev Rational Approximation
+        # 4.3 Chebyshev Rational Approximation
         delta_chebyshev_rational = np.abs(y_chebyshev_rational - plot_nodes_y_actual)
         print(F"Chebyshev Rational: {np.amax(delta_chebyshev_rational)}")
+
+    #############################
+    # 5. Plot Every Approximation
+
+    if shouldPlotApproximations:
+        # 5.1 Equispaced Polynomial Approximation
+        plt.plot(plot_nodes_x, y_eqispaced_poly, label="Equispaced Poly")
+
+        # 5.2 Chebyshev Polynomial Approximation
+        plt.plot(plot_nodes_x, y_chebyshev_poly, label="Chebyshev Poly")
+
+        # 5.3 Chebyshev Rational Approximation
+        plt.plot(plot_nodes_x, y_chebyshev_rational, label="Chebyshev Rational")
+
+        # 5.4 Actual With Large Number of Coordinates (evenly spaced on the X-axis)
+        plt.plot(plot_nodes_x, y_actual, label=F"Actual")
+
+        plt.legend(loc="upper left")
+        plt.grid(True)
+        plt.title(f"Appproximation of e^x on [{x_start}, {x_end}] with {num_parameters} parameters")
+        plt.show()
 
 if __name__ == "__main__":
     main()
